In em302_process.sh, replaced the use of mbgetesf with the mbcopy -M merge script.
Following a discussion with Gabriel Joyal, the decision was made to use mbcopy -M because it will not only merge the bathymetric edits, but also any processing perfomed in CARIS HIPS & SIPS: for example, a tidal reduction or a sound velocity correction. This is because (based on a discussion thread on the MB-System forum with Gordon Keith), the mbcopy -M option will extract the along-track, across-track and depth from the gsf file.